+++ pgjdbc/src/main/java/org/postgresql/ds/common/BaseDataSource.java
@@ -97,7 +97,7 @@ public abstract class BaseDataSource implements CommonDataSource, Referenceable 
       }
       return con;
     } catch (SQLException e) {
-      LOGGER.log(Level.SEVERE, "Failed to create a {0} for {1} at {2}: {3}",
+      LOGGER.log(Level.FINE, "Failed to create a {0} for {1} at {2}: {3}",
           new Object[]{getDescription(), user, getUrl(), e});
       throw e;
     }